Breaks HTMLInstrumentation into three files.
HTMLInstrumentation: handles the interface between editor/document
and live HTML.

HTMLSimpleDOM: Parses HTML into a simple DOM-like structure with
properties useful for HTMLInstrumentation.

HTMLDiff: Generates a list of edits to mutate a Simple DOMs into another.
