+++ src/language/HTMLSimpleDOM.js
@@ -1,0 +1,441 @@
+/*
+ * Copyright (c) 2013 Adobe Systems Incorporated. All rights reserved.
+ *  
+ * Permission is hereby granted, free of charge, to any person obtaining a
+ * copy of this software and associated documentation files (the "Software"), 
+ * to deal in the Software without restriction, including without limitation 
+ * the rights to use, copy, modify, merge, publish, distribute, sublicense, 
+ * and/or sell copies of the Software, and to permit persons to whom the 
+ * Software is furnished to do so, subject to the following conditions:
+ *  
+ * The above copyright notice and this permission notice shall be included in
+ * all copies or substantial portions of the Software.
+ *  
+ * THE SOFTWARE IS PROVIDED "AS IS", WITHOUT WARRANTY OF ANY KIND, EXPRESS OR
+ * IMPLIED, INCLUDING BUT NOT LIMITED TO THE WARRANTIES OF MERCHANTABILITY, 
+ * FITNESS FOR A PARTICULAR PURPOSE AND NONINFRINGEMENT. IN NO EVENT SHALL THE
+ * AUTHORS OR COPYRIGHT HOLDERS BE LIABLE FOR ANY CLAIM, DAMAGES OR OTHER 
+ * LIABILITY, WHETHER IN AN ACTION OF CONTRACT, TORT OR OTHERWISE, ARISING 
+ * FROM, OUT OF OR IN CONNECTION WITH THE SOFTWARE OR THE USE OR OTHER 
+ * DEALINGS IN THE SOFTWARE.
+ * 
+ */
+
+
+/*jslint vars: true, plusplus: true, devel: true, nomen: true, indent: 4, maxerr: 50 */
+/*global define, $, CodeMirror */
+/*unittests: HTML Instrumentation*/
+
+define(function (require, exports, module) {
+    "use strict";
+    
+    var DocumentManager = require("document/DocumentManager"),
+        Tokenizer       = require("language/HTMLTokenizer").Tokenizer,
+        MurmurHash3     = require("thirdparty/murmurhash3_gc"),
+        PerfUtils       = require("utils/PerfUtils");
+    
+    var seed = Math.floor(Math.random() * 65535);
+    
+    var tagID = 1;
+    
+    /**
+     * A list of tags whose start causes any of a given set of immediate parent
+     * tags to close. This mostly comes from the HTML5 spec section on omitted close tags:
+     * http://www.w3.org/html/wg/drafts/html/master/syntax.html#optional-tags
+     * This doesn't handle general content model violations.
+     */
+    var openImpliesClose = {
+        li      : { li: true },
+        dt      : { dd: true, dt: true },
+        dd      : { dd: true, dt: true },
+        address : { p: true },
+        article : { p: true },
+        aside   : { p: true },
+        blockquote : { p: true },
+        dir     : { p: true },
+        div     : { p: true },
+        dl      : { p: true },
+        fieldset: { p: true },
+        footer  : { p: true },
+        form    : { p: true },
+        h1      : { p: true },
+        h2      : { p: true },
+        h3      : { p: true },
+        h4      : { p: true },
+        h5      : { p: true },
+        h6      : { p: true },
+        header  : { p: true },
+        hgroup  : { p: true },
+        hr      : { p: true },
+        main    : { p: true },
+        menu    : { p: true },
+        nav     : { p: true },
+        ol      : { p: true },
+        p       : { p: true },
+        pre     : { p: true },
+        section : { p: true },
+        table   : { p: true },
+        ul      : { p: true },
+        rt      : { rp: true, rt: true },
+        rp      : { rp: true, rt: true },
+        optgroup: { optgroup: true, option: true },
+        option  : { option: true },
+        tbody   : { thead: true, tbody: true, tfoot: true },
+        tfoot   : { tbody: true },
+        tr      : { tr: true, th: true, td: true },
+        th      : { th: true, td: true },
+        td      : { thead: true, th: true, td: true },
+        body    : { head: true, link: true, script: true }
+    };
+
+    /**
+     * A list of tags that are self-closing (do not contain other elements).
+     * Mostly taken from http://www.w3.org/html/wg/drafts/html/master/syntax.html#void-elements
+     */
+    var voidElements = {
+        area: true,
+        base: true,
+        basefont: true,
+        br: true,
+        col: true,
+        command: true,
+        embed: true,
+        frame: true,
+        hr: true,
+        img: true,
+        input: true,
+        isindex: true,
+        keygen: true,
+        link: true,
+        menuitem: true,
+        meta: true,
+        param: true,
+        source: true,
+        track: true,
+        wbr: true
+    };
+    
+    function SimpleDOMBuilder(text, startOffset, startOffsetPos) {
+        this.stack = [];
+        this.text = text;
+        this.t = new Tokenizer(text);
+        this.currentTag = null;
+        this.startOffset = startOffset || 0;
+        this.startOffsetPos = startOffsetPos || {line: 0, ch: 0};
+    }
+    
+    function _updateHash(node) {
+        if (node.children) {
+            var i,
+                subtreeHashes = "",
+                childHashes = "",
+                child;
+            for (i = 0; i < node.children.length; i++) {
+                child = node.children[i];
+                if (child.children) {
+                    childHashes += String(child.tagID);
+                    subtreeHashes += String(child.tagID) + child.attributeSignature + child.subtreeSignature;
+                } else {
+                    childHashes += child.textSignature;
+                    subtreeHashes += child.textSignature;
+                }
+            }
+            node.childSignature = MurmurHash3.hashString(childHashes, childHashes.length, seed);
+            node.subtreeSignature = MurmurHash3.hashString(subtreeHashes, subtreeHashes.length, seed);
+        } else {
+            node.textSignature = MurmurHash3.hashString(node.content, node.content.length, seed);
+        }
+    }
+    
+    function _updateAttributeHash(node) {
+        var attributeString = JSON.stringify(node.attributes);
+        node.attributeSignature = MurmurHash3.hashString(attributeString, attributeString.length, seed);
+    }
+    
+    /**
+     * Generates a synthetic ID for text nodes. These IDs are only used
+     * for comparison purposes in the SimpleDOM structure, since we can't
+     * apply IDs to text nodes in the browser.
+     *
+     * TODO/Note: When generating a diff, the decision to do a textReplace 
+     * edit rather than a textDelete/textInsert hinges on how this ID
+     * is treated (because it is currently based on the previous or
+     * parent node).
+     *
+     * @param {Object} textNode new node for which we are generating an ID
+     * @return {string} ID for the node
+     */
+    function getTextNodeID(textNode) {
+        var childIndex = textNode.parent.children.indexOf(textNode);
+        if (childIndex === 0) {
+            return textNode.parent.tagID + ".0";
+        }
+        return textNode.parent.children[childIndex - 1].tagID + "t";
+    }
+    
+    function addPos(pos1, pos2) {
+        return {line: pos1.line + pos2.line, ch: (pos2.line === 0 ? pos1.ch + pos2.ch : pos2.ch)};
+    }
+    
+    function offsetPos(pos, offset) {
+        // Simple character offset. Only safe if the offset doesn't cross a line boundary.
+        return {line: pos.line, ch: pos.ch + offset};
+    }
+
+    SimpleDOMBuilder.prototype.build = function (strict) {
+        var self = this;
+        var token, lastClosedTag, lastTextNode, lastIndex = 0;
+        var stack = this.stack;
+        var attributeName = null;
+        var nodeMap = {};
+        var markCache = {};
+        
+        // Start timers for building full and partial DOMs.
+        // Appropriate timer is used, and the other is discarded.
+        var timerBuildFull = "HTMLInstr. Build DOM Full";
+        var timerBuildPart = "HTMLInstr. Build DOM Partial";
+        PerfUtils.markStart([timerBuildFull, timerBuildPart]);
+        
+        function closeTag(endIndex, endPos) {
+            lastClosedTag = stack[stack.length - 1];
+            stack.pop();
+            _updateHash(lastClosedTag);
+            
+            lastClosedTag.end = self.startOffset + endIndex;
+            lastClosedTag.endPos = addPos(self.startOffsetPos, endPos);
+        }
+        
+        while ((token = this.t.nextToken()) !== null) {
+            // lastTextNode is used to glue text nodes together
+            // If the last node we saw was text but this one is not, then we're done gluing.
+            // If this node is a comment, we might still encounter more text.
+            if (token.type !== "text" && token.type !== "comment" && lastTextNode) {
+                lastTextNode = null;
+            }
+            
+            if (token.type === "error") {
+                PerfUtils.finalizeMeasurement(timerBuildFull);  // discard
+                PerfUtils.addMeasurement(timerBuildPart);       // use
+                return null;
+            } else if (token.type === "opentagname") {
+                var newTagName = token.contents.toLowerCase(),
+                    newTag;
+                
+                if (openImpliesClose.hasOwnProperty(newTagName)) {
+                    var closable = openImpliesClose[newTagName];
+                    while (stack.length > 0 && closable.hasOwnProperty(stack[stack.length - 1].tag)) {
+                        // Close the previous tag at the start of this tag.
+                        // Adjust backwards for the < before the tag name.
+                        closeTag(token.start - 1, offsetPos(token.startPos, -1));
+                    }
+                }
+                
+                newTag = {
+                    tag: token.contents.toLowerCase(),
+                    children: [],
+                    attributes: {},
+                    parent: (stack.length ? stack[stack.length - 1] : null),
+                    start: this.startOffset + token.start - 1,
+                    startPos: addPos(this.startOffsetPos, offsetPos(token.startPos, -1)) // ok because we know the previous char was a "<"
+                };
+                newTag.tagID = this.getID(newTag, markCache);
+                
+                // During undo in particular, it's possible that tag IDs may be reused and
+                // the marks in the document may be misleading. If a tag ID has been reused,
+                // we apply a new tag ID to ensure that our edits come out correctly.
+                if (nodeMap[newTag.tagID]) {
+                    newTag.tagID = this.getNewID();
+                }
+                
+                nodeMap[newTag.tagID] = newTag;
+                if (newTag.parent) {
+                    newTag.parent.children.push(newTag);
+                }
+                this.currentTag = newTag;
+                
+                if (voidElements.hasOwnProperty(newTag.tag)) {
+                    // This is a self-closing element.
+                    _updateHash(newTag);
+                } else {
+                    stack.push(newTag);
+                }
+            } else if (token.type === "opentagend" || token.type === "selfclosingtag") {
+                // TODO: disallow <p/>?
+                if (this.currentTag) {
+                    // We're closing an open tag. Record the end of the open tag as the end of the
+                    // range. (If we later find a close tag for this tag, the end will get overwritten
+                    // with the end of the close tag. In the case of a self-closing tag, we should never
+                    // encounter that.)
+                    // Note that we don't need to update the signature here because the signature only
+                    // relies on the tag name and ID, and isn't affected by the tag's attributes, so
+                    // the signature we calculated when creating the tag is still the same. If we later
+                    // find a close tag for this tag, we'll update the signature to account for its
+                    // children at that point (in the next "else" case).
+                    this.currentTag.end = this.startOffset + token.end;
+                    this.currentTag.endPos = addPos(this.startOffsetPos, token.endPos);
+                    lastClosedTag = this.currentTag;
+                    _updateAttributeHash(this.currentTag);
+                    this.currentTag = null;
+                }
+            } else if (token.type === "closetag") {
+                // If this is a self-closing element, ignore the close tag.
+                var closeTagName = token.contents.toLowerCase();
+                if (!voidElements.hasOwnProperty(closeTagName)) {
+                    // Find the topmost item on the stack that matches. If we can't find one, assume
+                    // this is just a dangling closing tag and ignore it.
+                    var i;
+                    for (i = stack.length - 1; i >= 0; i--) {
+                        if (stack[i].tag === closeTagName) {
+                            break;
+                        }
+                    }
+                    if (strict && i !== stack.length - 1) {
+                        // If we're in strict mode, treat unbalanced tags as invalid.
+                        PerfUtils.finalizeMeasurement(timerBuildFull);
+                        PerfUtils.addMeasurement(timerBuildPart);
+                        return null;
+                    }
+                    if (i >= 0) {
+                        do {
+                            // For all tags we're implicitly closing (before we hit the matching tag), we want the
+                            // implied end to be the beginning of the close tag (which is two characters, "</", before
+                            // the start of the tagname). For the actual tag we're explicitly closing, we want the
+                            // implied end to be the end of the close tag (which is one character, ">", after the end of
+                            // the tagname).
+                            if (stack.length === i + 1) {
+                                closeTag(token.end + 1, offsetPos(token.endPos, 1));
+                            } else {
+                                closeTag(token.start - 2, offsetPos(token.startPos, -2));
+                            }
+                        } while (stack.length > i);
+                    } else {
+                        // If we're in strict mode, treat unmatched close tags as invalid. Otherwise
+                        // we just silently ignore them.
+                        if (strict) {
+                            PerfUtils.finalizeMeasurement(timerBuildFull);
+                            PerfUtils.addMeasurement(timerBuildPart);
+                            return null;
+                        }
+                    }
+                }
+            } else if (token.type === "attribname") {
+                attributeName = token.contents.toLowerCase();
+                // Set the value to the empty string in case this is an empty attribute. If it's not,
+                // it will get overwritten by the attribvalue later.
+                this.currentTag.attributes[attributeName] = "";
+            } else if (token.type === "attribvalue" && attributeName !== null) {
+                this.currentTag.attributes[attributeName] = token.contents;
+                attributeName = null;
+            } else if (token.type === "text") {
+                if (stack.length) {
+                    var parent = stack[stack.length - 1];
+                    var newNode;
+                    
+                    // Check to see if we're continuing a previous text.
+                    if (lastTextNode) {
+                        newNode = lastTextNode;
+                        newNode.content += token.contents;
+                    } else {
+                        newNode = {
+                            parent: stack[stack.length - 1],
+                            content: token.contents
+                        };
+                        parent.children.push(newNode);
+                        newNode.tagID = getTextNodeID(newNode);
+                        nodeMap[newNode.tagID] = newNode;
+                        lastTextNode = newNode;
+                    }
+                    
+                    _updateHash(newNode);
+                }
+            }
+            lastIndex = token.end;
+        }
+        
+        // If we have any tags hanging open (e.g. html or body), fail the parse if we're in strict mode,
+        // otherwise close them at the end of the document.
+        if (stack.length) {
+            if (strict) {
+                PerfUtils.finalizeMeasurement(timerBuildFull);
+                PerfUtils.addMeasurement(timerBuildPart);
+                return null;
+            } else {
+                // Manually compute the position of the end of the text (we can't rely on the
+                // tokenizer for this since it may not get to the very end)
+                // TODO: should probably make the tokenizer get to the end...
+                var lines = this.text.split("\n"),
+                    lastPos = {line: lines.length - 1, ch: lines[lines.length - 1].length};
+                while (stack.length) {
+                    closeTag(this.text.length, lastPos);
+                }
+            }
+        }
+        
+        var dom = lastClosedTag;
+        if (!dom) {
+            // This can happen if the document has no nontrivial content, or if the user tries to
+            // have something at the root other than the HTML tag. In all such cases, we treat the
+            // document as invalid.
+            return null;
+        }
+        
+        dom.nodeMap = nodeMap;
+        PerfUtils.addMeasurement(timerBuildFull);       // use
+        PerfUtils.finalizeMeasurement(timerBuildPart);  // discard
+        
+        return dom;
+    };
+    
+    /**
+     * Returns a new tag ID.
+     *
+     * @return {int} unique tag ID
+     */
+    SimpleDOMBuilder.prototype.getNewID = function () {
+        return tagID++;
+    };
+    
+    /**
+     * Returns the best tag ID for the new tag object given. 
+     * The default implementation just calls `getNewID`
+     * and returns a unique ID.
+     *
+     * @param {Object} newTag tag object to potentially inspect to choose an ID
+     * @return {int} unique tag ID
+     */
+    SimpleDOMBuilder.prototype.getID = SimpleDOMBuilder.prototype.getNewID;
+    
+    function buildSimpleDOM(text, strict) {
+        var builder = new SimpleDOMBuilder(text);
+        return builder.build(strict);
+    }
+    
+    function _dumpDOM(root) {
+        var result = "",
+            indent = "";
+        
+        function walk(node) {
+            if (node.tag) {
+                result += indent + "TAG " + node.tagID + " " + node.tag + " " + JSON.stringify(node.attributes) + "\n";
+            } else {
+                result += indent + "TEXT " + node.tagID + " " + node.content + "\n";
+            }
+            if (node.children) {
+                indent += "  ";
+                node.children.forEach(walk);
+                indent = indent.slice(2);
+            }
+        }
+        walk(root);
+        
+        return result;
+    }
+    
+    exports._dumpDOM                    = _dumpDOM;
+    exports.buildSimpleDOM              = buildSimpleDOM;
+    exports._offsetPos                  = offsetPos;
+    exports._updateHash                 = _updateHash;
+    exports._getTextNodeID              = getTextNodeID;
+    exports._seed                       = seed;
+    exports.SimpleDOMBuilder            = SimpleDOMBuilder;
+});
